Comments, no warnings, close #8
Also add pain header

+++ student-distrib/keyboard.c
@@ -17,7 +17,7 @@ void keyboard_init(){
 
 void keyboard_handler(){
     char byte = inb(KB_PORT);
-    char character = scan_code_1[byte];
+    char character = scan_code_1[(int) byte];
     putc(character);
     send_eoi(KB_IRQ);
 }
